# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DNDWORLDAPP API",
    "version": "0.1.0",
    "description": "Self-hosted worldbuilding and campaign API.\n\nTwo ways to authenticate: a `dwa_session` cookie for the browser, or\n`Authorization: Bearer <token>` for scripts and the MCP server.\n\nA token acts as its owner and inherits that user's role in each world.\nScopes and the optional world pin only ever *narrow* that — a token can\nnever grant more than the person holding it has.\n\nThings the caller may not see answer 404 rather than 403, so ids cannot\nbe probed to discover hidden content.\n\nThis document is generated from the Zod schemas in packages/schema."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerToken": []
    }
  ],
...
  "paths": {
...
    "/nodes/{nodeId}/map": {
      "get": {
        "tags": [
          "maps"
        ],
        "summary": "A page's map — source image, pixel bounds, and tiling status",
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Node id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "map": {
                      "$ref": "#/components/schemas/MapDto"
                    }
                  },
                  "required": [
                    "map"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the token is unknown, revoked or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed — including a token missing a scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Also returned for things the caller may not see, so ids cannot be probed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "maps"
        ],
        "summary": "Set or replace a page's map source image",
        "description": "The image must already be an uploaded asset in this world (see POST /worlds/{worldId}/assets). Replacing the image resets tiling.",
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Node id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMapImageInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "map": {
                      "$ref": "#/components/schemas/MapDto"
                    }
                  },
                  "required": [
                    "map"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the token is unknown, revoked or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed — including a token missing a scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Also returned for things the caller may not see, so ids cannot be probed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/nodes/{nodeId}/map/markers": {
      "get": {
        "tags": [
          "maps"
        ],
        "summary": "Markers on this map that the caller may see",
        "description": "DM-only markers are filtered in SQL, the same as DM fields and posts.",
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Node id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "markers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MapMarkerDto"
                      }
                    }
                  },
                  "required": [
                    "markers"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the token is unknown, revoked or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed — including a token missing a scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Also returned for things the caller may not see, so ids cannot be probed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "maps"
        ],
        "summary": "Add a marker to this map",
        "description": "A marker linked to a page (targetNodeId) shows that page's title/icon unless label/icon are set explicitly. Same edit gate as a field.",
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Node id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMarkerInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "marker": {
                      "$ref": "#/components/schemas/MapMarkerDto"
                    }
                  },
                  "required": [
                    "marker"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the token is unknown, revoked or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed — including a token missing a scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Also returned for things the caller may not see, so ids cannot be probed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/markers/{markerId}": {
      "patch": {
        "tags": [
          "maps"
        ],
        "summary": "Update a marker",
        "parameters": [
          {
            "name": "markerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Marker id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMarkerInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "marker": {
                      "$ref": "#/components/schemas/MapMarkerDto"
                    }
                  },
                  "required": [
                    "marker"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the token is unknown, revoked or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed — including a token missing a scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Also returned for things the caller may not see, so ids cannot be probed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "maps"
        ],
        "summary": "Delete a marker",
        "parameters": [
          {
            "name": "markerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Marker id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the token is unknown, revoked or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed — including a token missing a scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Also returned for things the caller may not see, so ids cannot be probed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MapDto": {
        "type": "object",
        "properties": {
          "nodeId": {
            "type": "string"
          },
          "assetUrl": {
            "type": "string"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "minZoom": {
            "type": "integer"
          },
          "maxZoom": {
            "type": "integer"
          },
          "tilingStatus": {
            "type": "string",
            "enum": [
              "none",
              "pending",
              "running",
              "ready",
              "error"
            ]
          },
          "fogEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "nodeId",
          "assetUrl",
          "width",
          "height",
          "minZoom",
          "maxZoom",
          "tilingStatus",
          "fogEnabled"
        ],
        "additionalProperties": false
      },
      "MapMarkerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "mapNodeId": {
            "type": "string"
          },
          "shape": {
            "type": "string",
            "enum": [
              "pin",
              "label",
              "circle",
              "polygon",
              "path",
              "token"
            ]
          },
          "x": {
            "type": "number"
          },
          "y": {
            "type": "number"
          },
          "points": {
            "type": "string",
            "nullable": true
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "members": {
            "type": "string",
            "nullable": true
          },
          "revealed": {
            "type": "boolean"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "members",
              "dm"
            ]
          },
          "targetNode": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "icon": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "title",
              "icon"
            ],
            "additionalProperties": false,
            "nullable": true
          },
          "parentMarkerId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "mapNodeId",
          "shape",
          "x",
          "y",
          "points",
          "label",
          "icon",
          "color",
          "members",
          "revealed",
          "visibility",
          "targetNode",
          "parentMarkerId"
        ],
        "additionalProperties": false
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "SetMapImageInput": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string"
          }
        },
        "required": [
          "assetId"
        ],
        "additionalProperties": false
      },
      "CreateMarkerInput": {
        "type": "object",
        "properties": {
          "shape": {
            "type": "string",
            "enum": [
              "pin",
              "label",
              "circle",
              "polygon",
              "path",
              "token"
            ],
            "default": "pin"
          },
          "x": {
            "type": "number"
          },
          "y": {
            "type": "number"
          },
          "points": {
            "type": "string",
            "maxLength": 20000
          },
          "targetNodeId": {
            "type": "string",
            "nullable": true
          },
          "label": {
            "type": "string",
            "maxLength": 300,
            "nullable": true
          },
          "icon": {
            "type": "string",
            "maxLength": 60,
            "nullable": true
          },
          "color": {
            "type": "string",
            "maxLength": 60,
            "nullable": true
          },
          "members": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "members",
              "dm"
            ],
            "default": "members"
          },
          "parentMarkerId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "x",
          "y"
        ],
        "additionalProperties": false
      },
      "UpdateMarkerInput": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number"
          },
          "y": {
            "type": "number"
          },
          "points": {
            "type": "string",
            "maxLength": 20000
          },
          "targetNodeId": {
            "type": "string",
            "nullable": true
          },
          "label": {
            "type": "string",
            "maxLength": 300,
            "nullable": true
          },
          "icon": {
            "type": "string",
            "maxLength": 60,
            "nullable": true
          },
          "color": {
            "type": "string",
            "maxLength": 60,
            "nullable": true
          },
          "members": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "revealed": {
            "type": "boolean"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "members",
              "dm"
            ]
          }
        },
        "additionalProperties": false
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dwa_session"
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A token from POST /tokens. Format: `dwa_` followed by 64 hex characters."
      }
    }
  }
}